requirements.txt
pillow
requests
python-Levenshtein